requirements.txt
django
djangorestframework
django-cors-headers
celery
psycopg
psycopg-binary
psycopg-pool
drf-yasg
django-environ